refactor(FE): move "FakemonView" component outside of Profile
Can be reused in Gym and Battle pages
