+++ frontend/src/components/FakemonView/FakemonView.js
@@ -1,3 +1,4 @@
+import React from "react";
 import styled from "@emotion/styled";
 
 function FakemonView({ fakemon }) {
