+++ frontend/src/components/FakemonView/index.js
@@ -1,0 +1,2 @@
+export * from './FakemonView';
+export { default } from './FakemonView';
